Websocket refactoring: define BackgroundThread
* Move sync background thread to new websocket submodule
* Split out the background thread functionality to a base class
